expert study html to image generation fix
There was an error while generating images from study.
When we were switching between 2 pages only containing scale or background nodes, the image generation was not triggered again.
This was due to the fact that the node IDs didn't change. We now use UUIDs as these two node IDs to avoid this.